requirements.txt
Django==1.6
Pillow==2.0.0
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.4.5
static==0.4
wsgiref==0.1.2
django-imagekit==3.0.4
South==0.8.4